+++ src/integration-test/kotlin/integration/chainadapter/ChainAdapterUnreadIntegrationTest.kt
@@ -87,7 +87,7 @@ class ChainAdapterUnreadIntegrationTest {
                 assertEquals(consumedTransactions.sorted(), consumedTransactions)
                 assertEquals(
                     adapter.getLastReadBlock(),
-                    adapter.lastReadBlockProvider.getLastBlockHeight()
+                    adapter.lastReadBlockProvider.getLastBlockHeight().toLong()
                 )
             }
         }
